Request pull request of your app only auth code
Hi Shogo! I see you've made an application only auth class: 
    http://shogo82148.github.io/blog/2013/05/09/application-only-authentication-with-tweepy/
Nice work!
    
Would you like to make it into a pull request to the main tweepy repo? It'd be good for it to be in there and attributed to you :)

I wasn't sure how else to ask you for it, so I'm doing it here (;
... though I managed to make a comment on your blog just now (wasn't sure what the email field was - my japanese is VERY rusty!)